fix: correct type for run_time variable and maintain consistency in prev_mouse_state initialization

+++ tests/dsp/dj_deck.vy
@@ -34,8 +34,8 @@ vaudio.attach_reverb(deck_b_track);
 vaudio.enable_eq(2);
 
 # --- INITIALIZE GLOBAL STATE VARIABLES ---
-run_time         :: Int64 = 0.0;
-prev_mouse_state :: Int64 = 0;
+run_time         :: Float64 = 0.0;
+prev_mouse_state :: Int64   = 0;
 
 # --- GLOBAL DJ MIXER STATE ---
 crossfader_pos :: Float64 = 0.5;   # 0.0 = Deck A Only, 1.0 = Deck B Only
